refactor: extract SVG exporter, rename data attributes, pass drawableArea as prop
- Extract downloadSvg into dedicated svgExporter.ts module
- Rename data-parent/data-child to data-erd-relation-parent-table-id/data-erd-relation-child-table-id
- Remove DRAWABLE_AREA export from support.ts, define in MainView and pass as prop through ControlPanel
- Update HTML export JS to use new data attribute names for perspective filtering
